Uninline JS and HTML (#126)
* Move indicator HTML to its own file

* move the shim too

* undo NeatInterpolation escapes

* Move the error message to its own file

* Have bin/run reload on ./www changes

* Update cabal file
