+++ www/ema-error.html
@@ -1,0 +1,15 @@
+<html>
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+</head>
+
+<body class="overflow-y: scroll;">
+    <h1>Ema App threw an exception</h1>
+    <pre
+        style="font-family: monospace; border: 1px solid; padding: 1em 1em 1em 1em; overflow-wrap: anywhere;">MESSAGE</pre>
+    <p>Once you fix the source of the error, this page will automatically refresh.
+</body>
+
+</html>
